Renamed qscaling.py to vmr_scaling.py.
