+++ CHANGELOG.txt
@@ -7514,4 +7514,9 @@ PEP8 cleaning extinction.py.
 
 PEP8 cleaning qscaling.py.
 
+*****
+
+Renamed qscaling.py to vmr_scaling.py.
+
+
