+++ pyratbay/atmosphere/vmr_scaling.py
